Change documentclass to article. Add to section 1.
On Tensors is unlikely to be long enough to justify usage of the book documentclass, so it was reverted to the article documentclass. A brief passage was added to address readers who are not familiar with fields. Another passage was added to better explain the reason for taking the dual basis to be the basis of the dual space.