Define |S| notation. Fix definition of linearity.
Section 1 employed the notation |S| to denote the cardinality of the set S without explaining it. This commit adds an explanation of this notation. The previous definition of linearity referred to mappings between vector spaces, when the actual mappings of concern are those from vector spaces to their fields. This commit fixes the definition of a linear mapping, making it refer to a mapping from a vector space to its field.